% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_indicators_municipios.R
\name{get_indicators_municipios}
\alias{get_indicators_municipios}
\title{Indicadores Municipais}
\usage{
get_indicators_municipios(indicator_name)
}
\arguments{
\item{indicator_name}{Um character com o nome do indicador municipal da plataforma
FGV - Municipios.}
}
\value{
Um character com o código do indicador.
}
\description{
Essa função retorna o código do indicador na API da plataforma FGV - Municípios
de acordo com o nome dos indicadores presentes na plataforma. O código do indicador
é usado na construção da requisição para a extração dos dados.
}
\examples{
\dontrun{
 # Pegando o código do indicador `Índice GINI`
  indicator_code <- get_indicators_municipios(indicator_name = "Índice GINI")
}
}
